feat(key_switch): 将按键相关代码独立出来，放到BSP中
